Add files via upload
How does a content-based recommendation system work?

Content-based filtering methods are based on a description of the item and a profile of the user's preferences. Content-based recommenders treat recommendation as a user-specific classification problem and learn a classifier for the user's likes and dislikes based on an item's features.

Basically, these methods use an item profile (i.e., a set of discrete attributes and characteristics) characterizing the item within the system. To abstract the features of the items in the system, an item presentation algorithm is applied. A widely used algorithm is the vector space representation.

Such a content-based recommendation system makes recommendations based on the "proximity" of an item's vector to the vector of an item that the user likes. Those items whose vector representations are close to the vector representation of an item already liked by the user are recommended.

Vector space model for recommendation systems

The Vector Space Model is used commonly in content-based recommendation systems. In this model, each item is represented as a vector of its attributes in an n-dimensional space and the angle between the vectors is calculated to determine the similarity between the vectors.

Cosine similarity measure is widely used to determine the proximity of two vectors represented using this model, and ultimately give a measure of the similarity between the vectors.